refactor: define getFile in a separate file
